Move uikit blocks
